requirements.txt
appdirs==1.4.3
asgiref==3.2.7
attrs==19.3.0
beautifulsoup4==4.9.0
black==19.10b0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Django==3.0.5
djangorestframework==3.11.0
idna==2.9
lxml==4.5.0
pathspec==0.8.0
pytz==2020.1
regex==2020.4.4
requests==2.23.0
socials==0.2.0
soupsieve==2.0
sqlparse==0.3.1
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.9
